+++ History.md
@@ -1,3 +1,17 @@
+0.1.3 - Apr 22, 2015
+====
+- initial support for web plafrom
+
+0.1.2 - Apr 19, 2015
+====
+- use yargs instead of commander
+- introduced local repo structure
+
+0.1.1 - Apr 18, 2015
+====
+- use config for main configuration
+- dir layout restructuring & application service refactoring
+
 0.1.0 - Jan 15, 2015
 ====
 - initial version